refactor(db): convert mock data seeding to factory pattern
This refactor reorganizes the mock data seeding utilities from standalone exported functions into a factory pattern with closure encapsulation. The change improves encapsulation by making internal helper functions truly private through closures rather than relying on export conventions.

Key changes:
- Created createMigrationTestData() factory that returns seedIndexedDB and clearIndexedDB methods
- Moved all internal helpers (generateMockAudio, generateMockRecording, etc.) into the factory's closure scope
- Updated MigrationDialog to instantiate and use the factory
- Fixed semantic HTML issues (changed non-form label elements to div elements)
- Added TypeScript typing for Switch component's onCheckedChange handler
- Added blog post documenting the factory pattern with closure encapsulation approach
